kservice: correctly type the unused legacy field as 8 bits
in 3f47342af76e4adf396100de768ea7e5fa2241a1 a qint8 was replaced with a
literal false because it was turned into an unused field. this is not a
compatible change because, to quote cppreference:

```
The value of sizeof(bool) is implementation defined and might differ
from 1.
```

this should not have had any negative effects though because AFAICT the
used streaming operator would cast to qint8 anyway. so, this mostly is
about the expression of our intent

+++ src/services/kservice.cpp
@@ -247,8 +247,8 @@ void KServicePrivate::save(QDataStream &s)
     // You may add new fields at the end. Make sure to update KSYCOCA_VERSION
     // number in ksycoca.cpp
     s << m_strType << m_strName << m_strExec << m_strIcon << term << m_strTerminalOptions << m_strWorkingDirectory << m_strComment
-      << false /* unused */ << m_mapProps << QString() /* unused */ << dst << m_strDesktopEntryName << m_lstKeywords << m_strGenName << categories << menuId
-      << m_actions << QStringList() /* unused */ << m_untranslatedName << m_untranslatedGenericName << m_mimeTypes;
+      << qint8(false) /* unused */ << m_mapProps << QString() /* unused */ << dst << m_strDesktopEntryName << m_lstKeywords << m_strGenName << categories
+      << menuId << m_actions << QStringList() /* unused */ << m_untranslatedName << m_untranslatedGenericName << m_mimeTypes;
 }
 
 ////
